% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_fits}
\alias{compare_fits}
\title{Compare fits between deSolve and macpan2}
\usage{
compare_fits(
  macpan2_results,
  deSolve_results,
  drop_last = TRUE,
  tolerance = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{macpan2_results}{results from running \code{report()} on a macpan2 simulator}

\item{deSolve_results}{results from \code{ode(...)} (typically want to use \code{method = "euler"} to match)}

\item{drop_last}{drop last point? (for reasons I don't yet understand, macpan2 diverges by a little bit
at last time point)}

\item{tolerance}{tolerance for numerical comparisons (set to 0 for exact comparison)}
}
\description{
Compare fits between deSolve and macpan2
}
\examples{
if (require("deSolve")) {
sirmod <- function(t, y, parameters) {
    grad <- with(as.list(c(y, parameters)),
                  c(- beta * S * I/N,
                    beta * S * I/N - gamma * I,
                     gamma * I))
    list(grad)
}

times <- 0:101
paras <- c(mu = 0, N = 100, gamma = 0.1, beta = 0.2)
start <- c(S = 0.999, I = 0.001, R = 0) * paras["N"]
deSolve_results = ode(y = start, times = times, func = sirmod, parms = paras,
                      method = "euler")

# macpan2
library(macpan2)

sir <- Compartmental(system.file("models", "sir_model_1", package = "macpan2helpers"))

sir_simulator = sir$simulators$tmb(
  time_steps = 100,
  state = start,
  flow = c(foi = NA_real_, gamma = paras[["gamma"]]),
  beta = paras[["beta"]],
  N = empty_matrix
)
macpan2_results = sir_simulator$report() 
comp <- compare_fits(macpan2_results, deSolve_results)
}
}
