% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawio.R
\name{drawio_tables}
\alias{drawio_tables}
\alias{drawio_table}
\title{Draw IO Table}
\usage{
drawio_tables(..., file_name = "")

drawio_table(
  data_frame,
  x = 0,
  y = 0,
  cell_width = 90,
  cell_height = 20,
  file_name = "",
  id_init = 2
)
}
\arguments{
\item{file_name}{Optional name of file. If missing, the xml will be returned.}

\item{data_frame}{The data frame to represent as a \code{draw.io} table.}

\item{x}{Horizontal position of the table.}

\item{y}{Vertical position of the table.}

\item{cell_width}{Width of table cells.}

\item{cell_height}{Height of table cells.}

\item{id_init}{Initial cell identifier. This is not usually need by users.}
}
\description{
Create a \code{draw.io} file with a table giving a data frame.
}
\section{Functions}{
\itemize{
\item \code{drawio_tables()}: Place several \code{draw.io} tables in a single file.
Each table is provided as a named list of the arguments to pass to
\code{drawio_table}.

}}
