% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_sort.R
\name{topological_sort}
\alias{topological_sort}
\title{Topological Sort}
\usage{
topological_sort(model, warn_not_dag = TRUE)
}
\arguments{
\item{model}{\code{\link{Compartmental}} model.}

\item{warn_not_dag}{Throw a warning if the model flows do not constitute
a directed acyclic graph (DAG)?}
}
\value{
Character vector giving the names of the state variables in
topologically sorted order.
}
\description{
Topological Sort
}
