% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visCompartmental.R
\name{visCompartmental}
\alias{visCompartmental}
\title{Visualize a Compartmental model}
\usage{
visCompartmental(model, label_flows = FALSE, ...)
}
\arguments{
\item{model}{Model object created using \code{\link[macpan2:Compartmental]{macpan2::Compartmental()}}}

\item{label_flows}{Logical. Should edges be labelled with flow rates?}

\item{...}{Additional arguments to pass to \code{\link{visNetwork}}}
}
\value{
A \code{visNetwork} object
}
\description{
Visualize a Compartmental model
}
\examples{
sir <- macpan2::Compartmental(system.file('starter_models', 'sir', package = 'macpan2'))
visCompartmental(sir)
}
