% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_fits}
\alias{compare_fits}
\title{Compare fits between \code{deSolve::ode()} and \code{macpan2}}
\usage{
compare_fits(
  macpan2_results,
  deSolve_results,
  drop_last = TRUE,
  tolerance = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{macpan2_results}{results from running \code{report()} on a macpan2 simulator}

\item{deSolve_results}{results from \code{ode(...)} (typically want to use \code{method = "euler"} to match)}

\item{drop_last}{drop last point? (for reasons I don't yet understand, macpan2 diverges by a little bit
at last time point)}

\item{tolerance}{tolerance for numerical comparisons (set to 0 for exact comparison)}
}
\value{
a list containing elements
\itemize{
\item \code{comb_plot}: combination plot (all variables)
\item \code{diff_plot}: difference plot
\item \code{all_equal}: results of \code{all.equal()}
\item \code{waldo_compare}: results of \code{waldo::compare} (if available)
}
}
\description{
Compare fits between \code{deSolve::ode()} and \code{macpan2}
}
