% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{mk_calibrate}
\alias{mk_calibrate}
\title{add calibration information to a simulatore}
\usage{
mk_calibrate(
  sim,
  params = list(),
  transforms = list(),
  data = list(),
  exprs = list()
)
}
\arguments{
\item{sim}{a \code{macpan2} simulator (i.e., a \code{TMBSimulator} object)}

\item{params}{a list of parameters with default/starting values}

\item{data}{a list-like object (list or data frame) containing data to add}

\item{exprs}{a list of expressions to add}
}
\description{
add calibration information to a simulatore
}
\examples{
m <- Compartmental(system.file("starter_models", "sir", package = "macpan2"))
sim <- m$simulators$tmb(
 time_steps = 100,
 state = c(S = 99, I = 1, R = 0),
 flow = c(foi = NA, gamma = 0.1),
 beta = 0.2,
 N = empty_matrix
 )
mk_calibrate(sim,
    data = list(I_obs = rep(0, 100)),
    params = list(beta = 0.2, I_sd = 1),
    transforms = list(beta = "Log", I_sd = "Log"),
    exprs = list(log_lik ~ dnorm(I_obs, I, I_sd))
)
}
